Move the Role model from app/models to templates/app/models.
Usually the application's app/models/ take precedence, but if
you use a shared gem for models across multiple applications...
it is possible for this gem's model to be loaded and your actual
model to be ignored.  Since this gem's model is barebones we can
put it as a template to be used instead.